Update Cliente double CPF

+++ Trabalho/cliente.h
@@ -13,27 +13,27 @@ using namespace std;
 class Cliente
 {
 private:
-    int cpf;
+    double cpf;
     string nome;
     int idade;
 public:
     Cliente(){}
-    Cliente(int cpf_, string nome_, int idade_) {
+    Cliente(double cpf_, string nome_, int idade_) {
         cpf = cpf_;
         nome = nome_;
         idade = idade_;
     }
-    Cliente(int cpf_, string nome_) {
+    Cliente(double cpf_, string nome_) {
         srand (time(NULL));
         cpf = cpf_;
         nome = nome_;
         idade = rand() % MAX_IDADE_RAND + MIN_IDADE_RAND;
     }
-    inline int getCpf(){return cpf;}
+    inline double getCpf(){return cpf;}
     inline string getNome(){return nome;}
     inline int getIdade(){return idade;}
 
-    inline void setCpf(int cpf_){cpf = cpf_;}
+    inline void setCpf(double cpf_){cpf = cpf_;}
     inline void setNome(string nome_){nome = nome_;}
     inline void setIdade(int idade_){idade = idade_;}
 };
